+++ HW1/do.sh
@@ -1,0 +1,4 @@
+#! /bin/sh
+flex hw1.l
+gcc lex.yy.c
+./a.out test.c
