+++ prac_1/analyze.c
