+++ prac_1/analyze.h
