+++ prac_1/globals.h
