+++ prac_1/scan.c
