+++ prac_1/scan.h
