+++ prac_1/systab.c
