+++ prac_1/systab.h
