+++ prac_1/t1.l
