+++ prac_1/tiny.l
