+++ prac_1/util.c
