+++ prac_1/util.h
